+++ inst/tinytest/test_fastshap_matrix.R
@@ -1,16 +1,24 @@
+# Exits
+if (!requireNamespace("earth", quietly = TRUE)) {
+  exit_file("Package earth missing")
+}
+if (!requireNamespace("ggplot2", quietly = TRUE)) {
+  exit_file("Package ggplot2 missing")
+}
+
 # Load required packages
-library(earth)    # for fitting MARS models
-library(ggplot2)
-library(mlbench)  # for ML benchmark data sets
+suppressMessages({
+  # library(earth)
+  library(ggplot2)
+})
 
 # Generate training data from the Friedman 1 benchmark problem
-set.seed(101)  # for reproducibility
-trn <- as.data.frame(mlbench.friedman1(500))
+trn <- gen_friedman(500, seed = 101)
 X_dframe <- subset(trn, select = -y)
 X_matrix <- data.matrix(X_dframe)
 
 # Fit a MARS model to the simulated Friedman benchmark data
-mars <- earth(y ~ ., data = trn, degree = 2)
+mars <- earth::earth(y ~ ., data = trn, degree = 2)
 
 # Prediction wrapper
 pfun <- function(object, newdata) {
@@ -32,8 +40,8 @@ expect_identical(
 # Check plots
 grid.arrange(
   autoplot(shap_dframe),
-  autoplot(shap_dframe, type = "dependence", X = X_dframe, feature = "x.3"),
+  autoplot(shap_dframe, type = "dependence", X = X_dframe, feature = "x3"),
   autoplot(shap_matrix),
-  autoplot(shap_matrix, type = "dependence", X = X_dframe, feature = "x.3"),
+  autoplot(shap_matrix, type = "dependence", X = X_dframe, feature = "x3"),
   nrow = 2
 )