+++ inst/tinytest/test_force_plot.R
@@ -9,17 +9,17 @@ if (!rstudioapi::isAvailable()) {
   exit_file("RStudio is not running.")
 }
 
-# Load required packages
-library(earth)    # for fitting MARS models
-library(mlbench)  # for ML benchmark data sets
+# # Load required packages
+# suppressMessages({
+#   library(earth)
+# })
 
 # Generate training data from the Friedman 1 benchmark problem
-set.seed(101)  # for reproducibility
-trn <- as.data.frame(mlbench.friedman1(500))
+trn <- gen_friedman(500, seed = 101)
 X <- subset(trn, select = -y)
 
 # Fit a MARS model to the simulated Friedman benchmark data
-mars <- earth(y ~ ., data = trn, degree = 2)
+mars <- eartth::earth(y ~ ., data = trn, degree = 2)
 
 # Prediction wrapper
 pfun <- function(object, newdata) {